+++ changelog.md
@@ -1,5 +1,9 @@
 # 1.0.3
+<<<<<<< HEAD
 - Added a back button to the Play menu, fixed endless scrolling, made level completion work better, and changed scene navigation.
+=======
+- Added a back button to the Play popup, fixed scrolling in Endless Mode, changed level completion detection, and fixed scene navigation.
+>>>>>>> cec3156 (Added more stuff, fixed bugs, fixed nitpicks, etc)
 
 # 1.0.1
 - Made it so that instead of fully deleting menu nodes, it instead makes them invisible. Turns out that's frowned upon.